moved to examples folder
decreased the amount of duplicate code
