+++ examples/automated_fetcher_csv.py
@@ -1,8 +1,7 @@
-#!/usr/bin/env python
+#!/usr/bin/env python2
 
 """
-Check if a web site returns a CloudFlare CAPTCHA using Tor Browser & httplib
-and submit results to an InfluxDB database
+Check if a web site returns a CloudFlare CAPTCHA using Tor Browser and httplib
 """
 
 import time
@@ -10,20 +9,15 @@ import sys
 import csv
 import itertools
 import os.path
-import json
-from influxdb import InfluxDBClient
 
+sys.path.append("..")
 import cloudflared_tor as cf_tor
 import cloudflared_httplib as cf_httplib
 
-host = 'localhost'
-port = 8086
-user_name = ''
-password = ''
-db = 'db_name'
-
 
 def main():
+    output_file = 'results.csv'
+
     url_list = ['http://captcha.wtf',
                 'https://captcha.wtf',
                 'http://captcha.wtf/complex.html',
@@ -50,39 +44,38 @@ def main():
 
     # Iterate over the url list
     for i, url in enumerate(url_list):
-        print('Testing %s' % url)
         params['url'] = url
 
         # Test with httplib
-        results = cf_httplib.is_cloudflared(params)
-        submit_to_influxdb(results)
+        test_with(cf_httplib, params, output_file)
 
         # Test with Tor
-        results = cf_tor.is_cloudflared(params)
-        submit_to_influxdb(results)
+        test_with(cf_tor, params, output_file)
+
+    print('> Completed testing...')
 
 
-# Submits given results to InfluxDB database
-def submit_to_influxdb(data):
-    # Connect to the database
-    client = InfluxDBClient(host, port, user_name, password, db)
+# Perform a test with the given paramters and append result to the CSV file
+def test_with(method, params, output_file):
+    results = method.is_cloudflared(params)
+    print('> Test result for %s with %s is %s' %
+            (results.get('url'), results.get('method'), results.get('result')))
+    append_to_csv(output_file, results)
 
-    # Prepare the json payload in the InfluxDB format
-    influxdb_json = [
-        {
-            'measurement': data['method'],
-            'tags': {
-                'url': data['url'],
-                'captcha_sign': data['captcha_sign']
-            },
-            'time': data['time_stamp'],
-            'fields': {
-                'result': data['result']
-            }
-        }
-    ]
 
-    client.write_points(influxdb_json)
+# Appends the passed result to the CSV file
+def append_to_csv(output_file, data):
+    # Create the output file if it doesn't exists
+    # Insert the header
+    if not os.path.isfile(output_file):
+        with open(output_file, 'w+') as file:
+            writer = csv.DictWriter(file, data.keys())
+            writer.writeheader()
+
+    # Append to file
+    with open(output_file, 'a') as file:
+        writer = csv.DictWriter(file, data.keys())
+        writer.writerow(data)
 
 
 if __name__ == '__main__':
